pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "strands-swarms"
version = "0.1.0"
description = "Task-based workflow orchestration for Strands Agents"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "Apache-2.0"}
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "strands-agents>=0.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=1.0.0",
    "ruff>=0.13.0",
    "mypy>=1.15.0",
]

[project.urls]
Homepage = "https://github.com/strands-agents/strands-swarms"
Issues = "https://github.com/strands-agents/strands-swarms/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/strands_swarms"]

[tool.ruff]
line-length = 120
include = ["src/**/*.py", "tests/**/*.py", "examples/**/*.py"]

[tool.ruff.lint]
select = ["E", "F", "I", "B"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_default_fixture_loop_scope = "function"
